Create BikeRental.java
Week 2- Exercise 1: Bike Rental
The aim of this exercise is to write a program that calculates bicycle rental amounts (for an
online 24/24 service).
The rental rates are as follows:
• From 00:00 to 07:00 and from 17:00 to 24:00: 1 Franc/hour
• From 07:00 to 17:00: 2 Francs/hour
The program will ask the user to enter the start and end hours of the rental as integers
(part of the code furnished) and will compute and display the rental cost.
You will adopt the following simplifications:
• the hours are integers (we don’t mind about minutes; each started hour will be fully
paid);
• the start time (i.e., hour) of the rental period must always be less than the end time
(i.e., hour) of the rent. This means that one cannot hire a bicycle more than 24 hours
and that each rental period must lie within the same day.
If the inputs are correct, then the program will calculate and show the rental amount on the
screen strictly following the format presented in the examples below. If the times entered
by the user are not between 0 and 24 (inclusive), then an error message strictly respecting
the following format will be shown and the program will terminate:
Les heures doivent être comprises entre 0 et 24 !
If the start time of the rent is equal to the end time, then an error message strictly
respecting the following format will be shown and the program will terminate:
Bizarre, vous n’avez pas loué votre vélo bien longtemps !
If the end time of the rent is less than the start time, then an error message strictly
respecting the following format will be shown and the program will terminate:
Bizarre, le début de la location est après la fin ...
The provided file already contains some instructions, that must be used as they are, in
order to produce the requested output. You must not use the instruction return; in this
program.