Create mostfrequent.java
EPFL-Week4 Assignement 1:
Most Frequent Element In An Erray
The purpose of this exercise is to write a program that identifies the most frequently occurring number in an array of integers. For example, for the following
array :
{2, 7, 5, 6, 7, 1, 6, 2, 1, 7}
your program is supposed to find that the most occurring number is 7 and that its
frequency is 3.