Create encryptation.java
EPFL-Week 4 Assignement:
Encryptation
The given program for this exercise provides you with a string ALPHABET containing all the letters of the latin alphabet, in lower case.
It asks the user to enter a string s.
The purpose of this exercise is to construct a new string aCoder as follows :
• aCoder contains, following their initial order, all the characters of s which
are present in ALPHABET or that correspond to the space character (’ ’).
All the other characters are ignored.
The string aCoder will be encoded using the technique of Julius Caesar. It has
to be displayed. If it does not contain any character, a message indicating that the
string to encode is empty will be displayed.
As long as the string to encode is non-empty, you program should encode it with
a fixed shifting given by the provided constant DECALAGE.
Encode the string aCoder by replacing every character with the character placed
DECALAGE letters further in the string ALPHABET, and this in a circular way (’z’
will be replaced by ’d’ for example).
The spaces will be maintained as they are. They are not object of any encoding
but will stay present at their position.
This encoded string will be displayed.